requirements.txt
requests
re
time
random
signal
sys
pwntools
